+++ R/precedence_matrix_length_two_loop.R
@@ -5,13 +5,9 @@
 #' @param eventlog The event log object to be used.
 #'
 #' @examples
-#' \dontrun{
-#' library(eventdataR)
-#' data(traffic_fines)
-#' m <- precedence_matrix_length_two_loop(traffic_fines)
+#' m <- precedence_matrix_length_two_loops(hospital_multi_perspective)
 #' print(m)
 #' as.matrix(m)
-#' }
 #'
 #' @export precedence_matrix_length_two_loops
 precedence_matrix_length_two_loops <- function(eventlog) {
@@ -27,7 +23,7 @@ precedence_matrix_length_two_loops_impl <- function(simplelog) {
 
   class(mat) <- c("process_matrix", class(mat))
   type <- "absolute"
-  attr(type, "perspective") <- "frequency"
+  attr(type, "perspective") <- "frequency_length_two_loops"
   attr(mat, "matrix_type") <- type
   return(mat)
 }